Move alert code to tendrl/commons
