+++ tendrl/commons/utils/command.py
@@ -2,6 +2,7 @@ import logging
 
 from ansible_module_runner import AnsibleExecutableGenerationFailed
 from ansible_module_runner import AnsibleRunner
+import shlex
 
 ANSIBLE_MODULE_PATH = "core/commands/command.py"
 LOG = logging.getLogger(__name__)
@@ -12,7 +13,8 @@ SAFE_COMMAND_LIST = [
     "lscpu",
     "getenforce",
     "gluster",
-    "ceph"
+    "ceph",
+    "config_manager"
 ]
 
 
@@ -24,7 +26,7 @@ class UnsupportedCommandException(Exception):
 
 class Command(object):
     def __init__(self, command):
-        if command.split()[0] not in SAFE_COMMAND_LIST:
+        if shlex.split(command)[0] not in SAFE_COMMAND_LIST:
             raise UnsupportedCommandException(command.split()[0])
         self.attributes = {"_raw_params": command}
